refactor: moving upgrade and update commands in sub-packages
fixes issue #4271
