+++ pkg/cmd/update/update_cluster_gke.go
@@ -1,4 +1,4 @@
-package cmd
+package update
 
 import (
 	"github.com/jenkins-x/jx/pkg/cloud"
