+++ pkg/cmd/upgrade/upgrade_apps.go
@@ -1,4 +1,4 @@
-package cmd
+package upgrade
 
 import (
 	"fmt"
