+++ pkg/cmd/upgrade/upgrade_cli_test.go
@@ -1,4 +1,4 @@
-package cmd
+package upgrade
 
 import (
 	"fmt"
