+++ pkg/cmd/upgrade/upgrade_cluster.go
@@ -1,4 +1,4 @@
-package cmd
+package upgrade
 
 import (
 	"encoding/json"
