+++ pkg/cmd/upgrade/upgrade_extensions.go
@@ -1,4 +1,4 @@
-package cmd
+package upgrade
 
 import (
 	"bytes"
